Expand delete-product-metafields-in-bulk to delete-product-or-product-variant-metafields-in-bulk
When dealing with product-related metatfields, merchants and developers often deal with product and product variant metafields. I updated the existing delete-product-metafields-in-bulk to handle cleanup for either products or product variant metafields.

I made a conscious decision to limit the task run to either product or product variant metafields. This decision ensures that the user knows which metafield namespace and key they are targeting (if they choose to supply those options).